Note push on save removed
Saving triggered a push to server. Previously the code was broken
so the push never happened. Now it's fixed, it's clear the push
happens in the UI thread, causing the app to freeze while it
takes place. This introduced an unacceptable delay during editing.

I've commented-out the push, since presumably things were working
okay when the push wasn't taking place. If it turns out the data
doesn't get pushed to the server at some other point, this should
be added back in, but with the push running in a separate thread.